SlotMap cleanup

+++ tst/packedmap_spec.c
@@ -274,7 +274,7 @@ describe(packedmap_foreach) {
 
   int input_data[] = { 0xAAAAAA, 0xBBBBB, 0xCCCC, 0xDDD, 0xEE, 0xF };
   PackedMap map = pmap_new(int);
-  for (int i = 0; i < ARRAY_COUNT(input_data); ++i) {
+  for (size_t i = 0; i < ARRAY_COUNT(input_data); ++i) {
     pmap_insert(map, &input_data[i]);
   }
 
@@ -309,7 +309,7 @@ describe(packedmap_views) {
   view_t expected = view_empty;
   int input_data[] = { 0xAAAAAA, 0xBBBBB, 0xCCCC, 0xDDD, 0xEE, 0xF };
   PackedMap map = pmap_new(int);
-  for (int i = 0; i < ARRAY_COUNT(input_data); ++i) {
+  for (size_t i = 0; i < ARRAY_COUNT(input_data); ++i) {
     pmap_insert(map, &input_data[i]);
   }
